test(bindings): lock down the RN kill-switch config parsing
The wire-format kill switches only work if the bridges actually
parse them, and that parsing lived in a private method on the
React module — untestable without a ReactApplicationContext, and
therefore untested. We already know how that story ends: four
encryption fields drifted to nested-only reads while the JS
wrapper sends them flat, and nobody noticed because nothing
*could* notice.

A kill switch that silently reverts to its default is not a kill
switch. It's a decoration.

Move the parse logic verbatim into ProtocolConfigParser (same
package, same read order) and pin the contract with JVM tests:
defaults when omitted, the flat camelCase shape the JS wrapper
actually sends, the nested encryption home, snake_case, and
nested-over-flat precedence. Constructing the UniFFI
ProtocolConfig in a plain JVM test is fine — it's a data class,
no native lib involved. iOS keeps its parseConfig inline (the
SwiftPM harness can't compile anything that imports the Generated
module), so it rides the keep-in-sync contract with Android.

While at it, fix the comment above the flag reads in both
bridges: it claimed "nested home first, then top level" for both
flags, but binaryWireEnabled has no nested home — top level IS
its home in the JS config and the flat UniFFI dictionary. Say so,
before someone helpfully "fixes" the asymmetry in either
direction.
